+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 Changelog
 =========
+### 4.3.13 - March 6, 2019 - hot fix release
+- Fixed the inability to accurately determine import section when a mix of conda and virtual environments are used.
+- Fixed some output being printed even when --quiet mode is enabled.
+- Fixed issue #890 interoperability with PyCharm by allowing case sensitive non type grouped sorting
+
+### 4.3.12 - March 6, 2019 - hot fix release
+- Fix error caused when virtual environment not detected
+
 ### 4.3.11 - March 6, 2019 - hot fix release
 - Fixed issue #876: confused by symlinks pointing to virtualenv gives FIRSTPARTY not THIRDPARTY
 - Fixed issue #873: current version skips every file on travis